20/02/2025

+++ src/main/java/org/dam19/tcgjpaserver/controllers/CartaController.java
@@ -18,26 +18,59 @@ public class CartaController {
     @PostMapping("/crear")
     public ResponseEntity<ResponseModel> crearCarta(@RequestBody CartaDto cartaDto,
                                                     @AuthenticationPrincipal UserDetails userDetails) {
-        boolean esAdmin = Boolean.valueOf(userDetails.getAuthorities().iterator().next().getAuthority());
-        if (esAdmin) {
+        if(userDetails == null){
+            return ResponseEntity.ok(new ResponseModel(1,"Usuario no autorizado",null));
+        }
+        String admin = userDetails.getAuthorities().iterator().next().getAuthority();
+        if (admin.equals("ROLE_true")) {
             return ResponseEntity.ok(cartaService.crearCarta(cartaDto));
         }
         return ResponseEntity.ok(new ResponseModel(1,"Usuario no autorizado",null));
     }
 
+
+
+
     @GetMapping("/buscar/{id}")
-    public ResponseEntity<ResponseModel> obtenerColeccionPorID(@PathVariable int id) {
-        return ResponseEntity.ok(cartaService.obtenerCartaPorId(id));
+    public ResponseEntity<ResponseModel> obtenerColeccionPorID(@PathVariable int id,@AuthenticationPrincipal UserDetails userDetails) {
+        if(userDetails == null){
+            return ResponseEntity.ok(new ResponseModel(1,"Usuario no autorizado",null));
+        }
+        String admin = userDetails.getAuthorities().iterator().next().getAuthority();
+        if (admin.equals("ROLE_true")) {
+            return ResponseEntity.ok(cartaService.obtenerCartaPorId(id));
+        }
+        return ResponseEntity.ok(new ResponseModel(1,"Usuario no autorizado",null));
     }
+
+
+
 
     @GetMapping("/todas")
-    public ResponseEntity<ResponseModel> obtenerTodasColecciones(){
-        return ResponseEntity.ok(cartaService.obtenerListaCartas());
+    public ResponseEntity<ResponseModel> obtenerTodasColecciones(@AuthenticationPrincipal UserDetails userDetails) {
+        if(userDetails == null){
+            return ResponseEntity.ok(new ResponseModel(1,"Usuario no autorizado",null));
+        }
+        String admin = userDetails.getAuthorities().iterator().next().getAuthority();
+        if (admin.equals("ROLE_true")) {
+            return ResponseEntity.ok(cartaService.obtenerListaCartas());
+        }
+        return ResponseEntity.ok(new ResponseModel(1,"Usuario no autorizado",null));
     }
 
+
+
+
     @GetMapping("/borrar/{id}")
-    public ResponseEntity<ResponseModel> eliminarCartaPorId(@PathVariable int id) {
-        return ResponseEntity.ok(cartaService.eliminarCartaPorId(id));
+    public ResponseEntity<ResponseModel> eliminarCartaPorId(@PathVariable int id,@AuthenticationPrincipal UserDetails userDetails) {
+        if(userDetails == null){
+            return ResponseEntity.ok(new ResponseModel(1,"Usuario no autorizado",null));
+        }
+        String admin = userDetails.getAuthorities().iterator().next().getAuthority();
+        if (admin.equals("ROLE_true")) {
+            return ResponseEntity.ok(cartaService.eliminarCartaPorId(id));
+        }
+       return ResponseEntity.ok(new ResponseModel(1,"Usuario no autorizado",null));
     }
 
 }
