+++ src/main/java/org/dam19/tcgjpaserver/controllers/ProductoController.java
@@ -6,6 +6,8 @@ import org.dam19.tcgjpaserver.models.ResponseModel;
 import org.dam19.tcgjpaserver.services.ProductoService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
@@ -15,23 +17,59 @@ public class ProductoController {
     private ProductoService productoService;
 
     @PostMapping("/crear")
-    public ResponseEntity<ResponseModel> crearProducto(@RequestBody ProductoDto productoDto) {
-        return ResponseEntity.ok(productoService.crearProducto(productoDto));
+    public ResponseEntity<ResponseModel> crearProducto(@RequestBody ProductoDto productoDto,
+                                                       @AuthenticationPrincipal UserDetails userDetails) {
+        if (userDetails == null) {
+            return ResponseEntity.ok(new ResponseModel(1,"Usuario no autorizado",null));
+        }
+        String admin = userDetails.getAuthorities().iterator().next().getAuthority();
+        if (admin.equals("ROLE_true")) {
+            return ResponseEntity.ok(productoService.crearProducto(productoDto));
+        }
+        return ResponseEntity.ok(new ResponseModel(1,"Usuario no autorizado",null));
     }
+
+
+
 
     @GetMapping("/buscar/{id}")
-    public ResponseEntity<ResponseModel> obtenerProductoPorID(@PathVariable int id) {
-        return ResponseEntity.ok(productoService.obtenerProductoPorId(id));
+    public ResponseEntity<ResponseModel> obtenerProductoPorID(@PathVariable int id,
+                                                              @AuthenticationPrincipal UserDetails userDetails) {
+        if (userDetails == null) {
+            return ResponseEntity.ok(new ResponseModel(1,"Usuario no autorizado",null));
+        }
+        String admin = userDetails.getAuthorities().iterator().next().getAuthority();
+        if (admin.equals("ROLE_true")) {
+            return ResponseEntity.ok(productoService.obtenerProductoPorId(id));
+        }
+        return ResponseEntity.ok(new ResponseModel(1,"Usuario no autorizado",null));
     }
+
+
 
     @GetMapping("/todos")
-    public ResponseEntity<ResponseModel> obtenerTodosProductos(){
-        return ResponseEntity.ok(productoService.obtenerListaProductos());
+    public ResponseEntity<ResponseModel> obtenerTodosProductos( @AuthenticationPrincipal UserDetails userDetails){
+        if (userDetails == null) {
+            return ResponseEntity.ok(new ResponseModel(1,"Usuario no autorizado",null));
+        }
+        String admin = userDetails.getAuthorities().iterator().next().getAuthority();
+        if (admin.equals("ROLE_true")) {
+            return ResponseEntity.ok(productoService.obtenerListaProductos());
+        }
+        return ResponseEntity.ok(new ResponseModel(1,"Usuario no autorizado",null));
     }
 
+
     @GetMapping("/borrar/{id}")
-    public ResponseEntity<ResponseModel> eliminaProductoPorId(@PathVariable int id) {
-        return ResponseEntity.ok(productoService.eliminarProductoPorId(id));
+    public ResponseEntity<ResponseModel> eliminaProductoPorId(@PathVariable int id, @AuthenticationPrincipal UserDetails userDetails) {
+        if (userDetails == null) {
+            return ResponseEntity.ok(new ResponseModel(1,"Usuario no autorizado",null));
+        }
+        String admin = userDetails.getAuthorities().iterator().next().getAuthority();
+        if (admin.equals("ROLE_true")) {
+            return ResponseEntity.ok(productoService.eliminarProductoPorId(id));
+        }
+       return ResponseEntity.ok(new ResponseModel(1,"Usuario no autorizado",null));
     }
 
 }
